+++ lib/marhan_cli/app/true_crypt_app.rb
@@ -3,22 +3,27 @@
 module MarhanCli
   class TrueCryptApp
 
-    @@binary = "/Applications/TrueCrypt.app/Contents/MacOS/TrueCrypt"
-
-    def initialize()
-      raise ArgumentError, "No installation of TrueCrypt found!" unless File.exist? @@binary
+    def initialize(mount_folder = nil)
+      @binary = "/Applications/TrueCrypt.app/Contents/MacOS/TrueCrypt"
+      raise ArgumentError, "No installation of TrueCrypt found!" unless File.exist? @binary
+      @mount_folder = mount_folder
     end
 
-    def mount_command()
-      "#{@@binary} --mount /dev/rdisk1s2 /Volumes/enc1"
+    def mount_command(encrypted_devices)
+      "#{@binary} --mount #{encrypted_devices[:enc1]} #{mount_folder}/#{:enc1.to_s}"
     end
 
     def unmount_command()
-      "#{@@binary} -d /Volumes/enc1"
+      "#{@binary} -d #{mount_folder}/enc1"
+    end
+
+    def mount_folder
+      raise ArgumentError, "No mount folder defined! Don't know where to mount!" unless @mount_folder
+      @mount_folder
     end
 
     def unmount_all_command()
-      "#{@@binary} -d"
+      "#{@binary} -d"
     end
 
   end